// apps/web/src/api/models/WaitlistResponseDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type WaitlistResponseDto = {
    id: string;
    userId: string;
    firstName: string;
    lastName: string;
    email: string;
    phone: string;
    allowed: boolean;
    allowedOn?: string;
    createdAt: string;
    updatedAt: string;
};